@@ -45,7 +45,8 @@
         "view:web:profile": "gulp view:web:profile",
         "bench": "npm run build:libs && lerna exec --scope @casual-simulation/aux-benchmarks -- npm run bench",
         "bench:profile": "npm run build:libs && lerna exec --scope @casual-simulation/aux-benchmarks -- npm run bench:profile",
-        "cloc": "cloc src --exclude-dir=node_modules --exclude-ext=js,json,def"
+        "cloc": "cloc src --exclude-dir=node_modules --exclude-ext=js,json,def",
+        "import-typescript": "lerna exec --scope @casual-simulation/aux-custom-portals -- npm run import-typescript"
     },
     "husky": {
         "hooks": {
@@ -132,7 +133,7 @@
         "ts-jest": "26.4.4",
         "ts-loader": "8.0.11",
         "typedoc": "^0.15.0",
-        "typescript": "^4.1.2",
+        "typescript": "^4.2.2",
         "vue-loader": "^15.9.5",
         "vue-svg-loader": "^0.16.0",
         "vue-template-compiler": "^2.5.17",
@@ -148,7 +149,8 @@
         "@types/faker": "5.1.3",
         "webpack-assets-manifest": "5.0.1",
         "@testing-library/vue": "5.6.1",
-        "@testing-library/jest-dom": "5.11.9"
+        "@testing-library/jest-dom": "5.11.9",
+        "@casual-simulation/typescript": "^4.2.4"
     },
     "dependencies": {}
 }
