@@ -51,7 +51,7 @@
         "rxjs": "^6.5.2"
     },
     "devDependencies": {
-        "typescript": "^4.1.2",
+        "typescript": "^4.2.2",
         "webpack": "^5.10.0",
         "webpack-cli": "^4.2.0",
         "webpack-merge": "^5.4.0",
