@@ -71,6 +71,8 @@ import {
     DNA_TAG_PREFIX,
     tagsOnBot,
     TEMPORARY_BOT_PARTITION_ID,
+    openCustomPortal,
+    registerBuiltinPortal,
 } from '../bots';
 import { v4 as uuid } from 'uuid';
 import { waitAsync } from '../test/TestHelpers';
@@ -552,11 +554,11 @@ describe('AuxRuntime', () => {
                         stateUpdatedEvent({
                             test1: createBot('test1', {
                                 abc: 'def',
-                                onBotAdded: `@player.toast("Added 1!")`,
+                                onBotAdded: `@os.toast("Added 1!")`,
                             }),
                             test2: createBot('test2', {
                                 abc: 'ghi',
-                                onBotAdded: `@player.toast("Added 2!")`,
+                                onBotAdded: `@os.toast("Added 2!")`,
                             }),
                             test3: createBot('test3', {
                                 abc: '999',
@@ -576,11 +578,11 @@ describe('AuxRuntime', () => {
                         stateUpdatedEvent({
                             test1: createBot('test1', {
                                 abc: 'def',
-                                onBotAdded: `@player.toast(getBots('abc', 'ghi').length + 10)`,
+                                onBotAdded: `@os.toast(getBots('abc', 'ghi').length + 10)`,
                             }),
                             test2: createBot('test2', {
                                 abc: 'ghi',
-                                onBotAdded: `@player.toast(getBots('abc', 'def').length + 20)`,
+                                onBotAdded: `@os.toast(getBots('abc', 'def').length + 20)`,
                             }),
                         })
                     );
@@ -595,7 +597,7 @@ describe('AuxRuntime', () => {
                         stateUpdatedEvent({
                             test1: createBot('test1', {
                                 abc: 'def',
-                                onBotAdded: `@player.toast(that)`,
+                                onBotAdded: `@os.toast(that)`,
                             }),
                         })
                     );
@@ -612,7 +614,7 @@ describe('AuxRuntime', () => {
                             test1: createBot('test1', {
                                 abc: 'def',
                                 create: `@create({
-                                onBotAdded: '@player.toast("hit")'
+                                onBotAdded: '@os.toast("hit")'
                             })`,
                             }),
                         })
@@ -627,7 +629,7 @@ describe('AuxRuntime', () => {
                             botAdded(
                                 createBot('uuid1', {
                                     creator: 'test1',
-                                    onBotAdded: '@player.toast("hit")',
+                                    onBotAdded: '@os.toast("hit")',
                                 })
                             ),
                         ],
@@ -640,11 +642,11 @@ describe('AuxRuntime', () => {
                         stateUpdatedEvent({
                             test1: createBot('test1', {
                                 abc: 'def',
-                                onBotAdded: `@player.toast("Added 1!")`,
+                                onBotAdded: `@os.toast("Added 1!")`,
                             }),
                             test2: createBot('test2', {
                                 abc: 'ghi',
-                                onBotAdded: `@player.toast("Added 2!")`,
+                                onBotAdded: `@os.toast("Added 2!")`,
                             }),
                             test3: createBot('test3', {
                                 abc: '999',
@@ -663,11 +665,11 @@ describe('AuxRuntime', () => {
                         stateUpdatedEvent({
                             test1: createBot('test1', {
                                 abc: 'def',
-                                onBotAdded: `@player.toast("Added 1!")`,
+                                onBotAdded: `@os.toast("Added 1!")`,
                             }),
                             test2: createBot('test2', {
                                 abc: 'ghi',
-                                onBotAdded: `@player.toast("Added 2!")`,
+                                onBotAdded: `@os.toast("Added 2!")`,
                             }),
                             test3: createBot('test3', {
                                 abc: '999',
@@ -687,7 +689,7 @@ describe('AuxRuntime', () => {
                         stateUpdatedEvent({
                             test1: createBot('test1', {
                                 abc: 'def',
-                                onAnyBotsAdded: `@player.toast(that.bots.length)`,
+                                onAnyBotsAdded: `@os.toast(that.bots.length)`,
                             }),
                             test2: createBot('test2', {
                                 abc: 'ghi',
@@ -708,7 +710,7 @@ describe('AuxRuntime', () => {
                         stateUpdatedEvent({
                             test1: createBot('test1', {
                                 abc: 'def',
-                                onAnyBotsAdded: `@player.toast(that.bots.length)`,
+                                onAnyBotsAdded: `@os.toast(that.bots.length)`,
                             }),
                         })
                     );
@@ -779,11 +781,11 @@ describe('AuxRuntime', () => {
                         stateUpdatedEvent({
                             test1: createBot('test1', {
                                 abc: 'def',
-                                onAnyBotsAdded: `@player.toast("Added 1!")`,
+                                onAnyBotsAdded: `@os.toast("Added 1!")`,
                             }),
                             test2: createBot('test2', {
                                 abc: 'ghi',
-                                onAnyBotsAdded: `@player.toast("Added 2!")`,
+                                onAnyBotsAdded: `@os.toast("Added 2!")`,
                             }),
                             test3: createBot('test3', {
                                 abc: '999',
@@ -802,11 +804,11 @@ describe('AuxRuntime', () => {
                         stateUpdatedEvent({
                             test1: createBot('test1', {
                                 abc: 'def',
-                                onAnyBotsAdded: `@player.toast("Added 1!")`,
+                                onAnyBotsAdded: `@os.toast("Added 1!")`,
                             }),
                             test2: createBot('test2', {
                                 abc: 'ghi',
-                                onAnyBotsAdded: `@player.toast("Added 2!")`,
+                                onAnyBotsAdded: `@os.toast("Added 2!")`,
                             }),
                             test3: createBot('test3', {
                                 abc: '999',
@@ -963,8 +965,7 @@ describe('AuxRuntime', () => {
                     const update1 = runtime.stateUpdated(
                         stateUpdatedEvent({
                             test: createBot('test', {
-                                abc:
-                                    '@setInterval(() => player.toast("hi"), 100);',
+                                abc: '@setInterval(() => os.toast("hi"), 100);',
                             }),
                         })
                     );
@@ -1069,7 +1070,7 @@ describe('AuxRuntime', () => {
                             }),
                             test3: createBot('test3', {
                                 abc: '999',
-                                onAnyBotsRemoved: `@player.toast(that.botIDs)`,
+                                onAnyBotsRemoved: `@os.toast(that.botIDs)`,
                             }),
                         })
                     );
@@ -1094,11 +1095,11 @@ describe('AuxRuntime', () => {
                             }),
                             test2: createBot('test2', {
                                 abc: 'ghi',
-                                onAnyBotsRemoved: `@player.toast(getBots('abc').length)`,
+                                onAnyBotsRemoved: `@os.toast(getBots('abc').length)`,
                             }),
                             test3: createBot('test3', {
                                 abc: '999',
-                                onAnyBotsRemoved: `@player.toast(getBots('abc').length + 10)`,
+                                onAnyBotsRemoved: `@os.toast(getBots('abc').length + 10)`,
                             }),
                         })
                     );
@@ -1122,7 +1123,7 @@ describe('AuxRuntime', () => {
                             test1: createBot('test1', {
                                 abc: 'def',
                                 destroy: `@destroy(this)`,
-                                onBotDestroyed: `@player.toast("Hit")`,
+                                onBotDestroyed: `@os.toast("Hit")`,
                             }),
                         })
                     );
@@ -1145,7 +1146,7 @@ describe('AuxRuntime', () => {
                             }),
                             test3: createBot('test3', {
                                 abc: '999',
-                                onAnyBotsRemoved: `@player.toast(that.botIDs)`,
+                                onAnyBotsRemoved: `@os.toast(that.botIDs)`,
                             }),
                         })
                     );
@@ -1174,7 +1175,7 @@ describe('AuxRuntime', () => {
                             }),
                             test3: createBot('test3', {
                                 abc: '999',
-                                onAnyBotsRemoved: `@player.toast(that.botIDs)`,
+                                onAnyBotsRemoved: `@os.toast(that.botIDs)`,
                             }),
                         })
                     );
@@ -1579,11 +1580,11 @@ describe('AuxRuntime', () => {
                         stateUpdatedEvent({
                             test1: createBot('test1', {
                                 abc: 'def',
-                                onBotChanged: `@player.toast("Changed 1!")`,
+                                onBotChanged: `@os.toast("Changed 1!")`,
                             }),
                             test2: createBot('test2', {
                                 abc: 'ghi',
-                                onBotChanged: `@player.toast("Changed 2!")`,
+                                onBotChanged: `@os.toast("Changed 2!")`,
                             }),
                             test3: createBot('test3', {
                                 abc: '999',
@@ -1618,7 +1619,7 @@ describe('AuxRuntime', () => {
                         stateUpdatedEvent({
                             test1: createBot('test1', {
                                 abc: 'def',
-                                onBotChanged: `@player.toast(getBots('zzz').length)`,
+                                onBotChanged: `@os.toast(getBots('zzz').length)`,
                             }),
                         })
                     );
@@ -1643,7 +1644,7 @@ describe('AuxRuntime', () => {
                         stateUpdatedEvent({
                             test1: createBot('test1', {
                                 abc: 'def',
-                                onBotChanged: `@player.toast(that)`,
+                                onBotChanged: `@os.toast(that)`,
                             }),
                         })
                     );
@@ -1675,11 +1676,11 @@ describe('AuxRuntime', () => {
                         stateUpdatedEvent({
                             test1: createBot('test1', {
                                 abc: 'def',
-                                onBotChanged: `@player.toast("Changed 1!")`,
+                                onBotChanged: `@os.toast("Changed 1!")`,
                             }),
                             test2: createBot('test2', {
                                 abc: 'ghi',
-                                onBotChanged: `@player.toast("Changed 2!")`,
+                                onBotChanged: `@os.toast("Changed 2!")`,
                             }),
                             test3: createBot('test3', {
                                 abc: '999',
@@ -1715,11 +1716,11 @@ describe('AuxRuntime', () => {
                         stateUpdatedEvent({
                             test1: createBot('test1', {
                                 abc: 'def',
-                                onBotChanged: `@player.toast("Changed 1!")`,
+                                onBotChanged: `@os.toast("Changed 1!")`,
                             }),
                             test2: createBot('test2', {
                                 abc: 'ghi',
-                                onBotChanged: `@player.toast("Changed 2!")`,
+                                onBotChanged: `@os.toast("Changed 2!")`,
                             }),
                             test3: createBot('test3', {
                                 abc: '999',
@@ -1757,7 +1758,7 @@ describe('AuxRuntime', () => {
                         stateUpdatedEvent({
                             test1: createBot('test1', {
                                 abc: 'def',
-                                onAnyBotsChanged: `@player.toast(that)`,
+                                onAnyBotsChanged: `@os.toast(that)`,
                             }),
                             test2: createBot('test2', {
                                 abc: 'ghi',
@@ -1850,11 +1851,11 @@ describe('AuxRuntime', () => {
                         stateUpdatedEvent({
                             test1: createBot('test1', {
                                 abc: 'def',
-                                onAnyBotsChanged: `@player.toast("Changed 1!")`,
+                                onAnyBotsChanged: `@os.toast("Changed 1!")`,
                             }),
                             test2: createBot('test2', {
                                 abc: 'ghi',
-                                onAnyBotsChanged: `@player.toast("Changed 2!")`,
+                                onAnyBotsChanged: `@os.toast("Changed 2!")`,
                             }),
                             test3: createBot('test3', {
                                 abc: '999',
@@ -1890,11 +1891,11 @@ describe('AuxRuntime', () => {
                         stateUpdatedEvent({
                             test1: createBot('test1', {
                                 abc: 'def',
-                                onAnyBotsChanged: `@player.toast("Changed 1!")`,
+                                onAnyBotsChanged: `@os.toast("Changed 1!")`,
                             }),
                             test2: createBot('test2', {
                                 abc: 'ghi',
-                                onAnyBotsChanged: `@player.toast("Changed 2!")`,
+                                onAnyBotsChanged: `@os.toast("Changed 2!")`,
                             }),
                             test3: createBot('test3', {
                                 abc: '999',
@@ -2929,10 +2930,10 @@ describe('AuxRuntime', () => {
             runtime.stateUpdated(
                 stateUpdatedEvent({
                     test1: createBot('test1', {
-                        hello: '@player.toast("hi1")',
+                        hello: '@os.toast("hi1")',
                     }),
                     test2: createBot('test2', {
-                        hello: '@player.toast("hi2")',
+                        hello: '@os.toast("hi2")',
                     }),
                     test3: createBot('test3', {}),
                 })
@@ -2948,10 +2949,10 @@ describe('AuxRuntime', () => {
             runtime.stateUpdated(
                 stateUpdatedEvent({
                     test1: createBot('test1', {
-                        hello: '@player.toast("hi1")',
+                        hello: '@os.toast("hi1")',
                     }),
                     test2: createBot('test2', {
-                        hello: '@player.toast("hi2")',
+                        hello: '@os.toast("hi2")',
                     }),
                     test3: createBot('test3', {}),
                 })
@@ -2968,7 +2969,7 @@ describe('AuxRuntime', () => {
         it('should flatten run script events into the given batch', async () => {
             runtime.process([
                 toast('hi0'),
-                runScript('player.toast("hi1")'),
+                runScript('os.toast("hi1")'),
                 toast('hi2'),
             ]);
 
@@ -2983,13 +2984,13 @@ describe('AuxRuntime', () => {
             runtime.stateUpdated(
                 stateUpdatedEvent({
                     test1: createBot('test1', {
-                        onServerAction: '@player.toast(that.action.message)',
+                        onServerAction: '@os.toast(that.action.message)',
                     }),
                 })
             );
             runtime.process([
                 toast('hi0'),
-                runScript('player.toast("hi1")'),
+                runScript('os.toast("hi1")'),
                 toast('hi2'),
             ]);
 
@@ -3018,7 +3019,7 @@ describe('AuxRuntime', () => {
             );
             runtime.process([
                 toast('hi0'),
-                runScript('player.toast("hi1")'),
+                runScript('os.toast("hi1")'),
                 toast('hi2'),
             ]);
 
@@ -3032,7 +3033,7 @@ describe('AuxRuntime', () => {
                 stateUpdatedEvent({
                     test1: createBot('test1', {
                         onServerAction: '@tags.count += 1',
-                        wow: '@player.toast("hi")',
+                        wow: '@os.toast("hi")',
                         count: 0,
                     }),
                 })
@@ -3040,7 +3041,7 @@ describe('AuxRuntime', () => {
             runtime.process([
                 toast('hi0'),
                 action('wow'),
-                runScript('player.toast("hi1")'),
+                runScript('os.toast("hi1")'),
                 toast('hi2'),
             ]);
 
@@ -3096,7 +3097,7 @@ describe('AuxRuntime', () => {
                 stateUpdatedEvent({
                     test1: createBot('test1', {
                         onServerAction: `@if(that.action.type === "action") action.reject(that.action);`,
-                        test: '@player.toast("hi")',
+                        test: '@os.toast("hi")',
                     }),
                 })
             );
@@ -3115,7 +3116,7 @@ describe('AuxRuntime', () => {
                     }),
                 })
             );
-            runtime.process([runScript('player.toast("hi")')]);
+            runtime.process([runScript('os.toast("hi")')]);
 
             await waitAsync();
 
@@ -3175,7 +3176,7 @@ describe('AuxRuntime', () => {
                     }),
                 })
             );
-            runtime.process([device(<any>{}, runScript('player.toast("hi")'))]);
+            runtime.process([device(<any>{}, runScript('os.toast("hi")'))]);
 
             await waitAsync();
 
@@ -3184,16 +3185,14 @@ describe('AuxRuntime', () => {
                     // onServerAction is executed before
                     // the device action is executed
                     toast('hi'),
-                    device(<any>{}, runScript('player.toast("hi")')),
+                    device(<any>{}, runScript('os.toast("hi")')),
                 ],
             ]);
         });
 
         it('should support resolving async actions', async () => {
             runtime.process([
-                runScript(
-                    'player.showInput().then(result => player.toast(result))'
-                ),
+                runScript('os.showInput().then(result => os.toast(result))'),
             ]);
 
             await waitAsync();
@@ -3222,7 +3221,7 @@ describe('AuxRuntime', () => {
 
             runtime.process([
                 runScript(
-                    'player.showInput().then(result => player.toast(result.tags.abc))'
+                    'os.showInput().then(result => os.toast(result.tags.abc))'
                 ),
             ]);
 
@@ -3252,9 +3251,7 @@ describe('AuxRuntime', () => {
 
         it('should support rejecting async actions', async () => {
             runtime.process([
-                runScript(
-                    'player.showInput().catch(result => player.toast(result))'
-                ),
+                runScript('os.showInput().catch(result => os.toast(result))'),
             ]);
 
             await waitAsync();
@@ -3276,7 +3273,7 @@ describe('AuxRuntime', () => {
             uuidMock.mockReturnValueOnce('task1');
             runtime.process([
                 runScript(
-                    'server.serverPlayerCount("test").then(result => player.toast(result))'
+                    'server.serverPlayerCount("test").then(result => os.toast(result))'
                 ),
             ]);
 
@@ -3297,7 +3294,7 @@ describe('AuxRuntime', () => {
             uuidMock.mockReturnValueOnce('task1');
             runtime.process([
                 runScript(
-                    'server.serverPlayerCount("test").catch(err => player.toast(err))'
+                    'server.serverPlayerCount("test").catch(err => os.toast(err))'
                 ),
             ]);
 
@@ -3317,8 +3314,8 @@ describe('AuxRuntime', () => {
         it('should support using await for async actions', async () => {
             runtime.process([
                 runScript(
-                    `const result = await player.showInput();
-                     player.toast(result);`
+                    `const result = await os.showInput();
+                     os.toast(result);`
                 ),
             ]);
 
@@ -3338,7 +3335,7 @@ describe('AuxRuntime', () => {
         });
 
         it('should not crash if given a run_script that doesnt compile', async () => {
-            runtime.process([runScript('player.toast('), toast('abc')]);
+            runtime.process([runScript('os.toast('), toast('abc')]);
 
             await waitAsync();
 
@@ -3347,7 +3344,7 @@ describe('AuxRuntime', () => {
 
         it('should resolve run_script tasks', async () => {
             const result = runtime.execute(
-                'return await player.run("return 123");'
+                'return await os.run("return 123");'
             );
 
             runtime.process(result.actions);
@@ -3357,18 +3354,189 @@ describe('AuxRuntime', () => {
 
         it('should unwrap async run_script tasks', async () => {
             const result = runtime.execute(
-                'return await player.run("return Promise.resolve(123);");'
+                'return await os.run("return Promise.resolve(123);");'
             );
 
             runtime.process(result.actions);
 
             expect(await result.result).toBe(123);
         });
+
+        describe('register_builtin_portal', () => {
+            it('should add a global variable with a new tempLocal bot for the bot included in the action', async () => {
+                uuidMock.mockReturnValueOnce('uuid');
+                runtime.process([registerBuiltinPortal('page')]);
+
+                await waitAsync();
+
+                expect(allEvents).toEqual([
+                    openCustomPortal('page', 'uuid', null, {}),
+                    botAdded(createBot('uuid', {}, 'tempLocal')),
+                ]);
+
+                expect((<any>globalThis).pageBot).toBe(
+                    runtime.context.state['uuid']
+                );
+            });
+
+            it('should not override previous variables', async () => {
+                runtime.stateUpdated(
+                    stateUpdatedEvent({
+                        test1: createBot('test1', {
+                            abc: 'def',
+                        }),
+                        test2: createBot('test2', {
+                            abc: 'other',
+                        }),
+                    })
+                );
+                runtime.process([
+                    openCustomPortal('page', 'test1', 'myTag', {}),
+                ]);
+
+                expect((<any>globalThis).pageBot).toBe(
+                    runtime.context.state['test1']
+                );
+
+                runtime.process([registerBuiltinPortal('page')]);
+
+                await waitAsync();
+
+                expect(allEvents).toEqual([
+                    openCustomPortal('page', 'test1', 'myTag', {}),
+                ]);
+                expect((<any>globalThis).pageBot).toBe(
+                    runtime.context.state['test1']
+                );
+            });
+
+            it('should remove the global variables that were created by the runtime', () => {
+                uuidMock.mockReturnValueOnce('uuid');
+
+                runtime.process([registerBuiltinPortal('page')]);
+
+                expect((<any>globalThis).pageBot).toBe(
+                    runtime.context.state['uuid']
+                );
+
+                runtime.unsubscribe();
+
+                expect(
+                    Object.getOwnPropertyDescriptor(globalThis, 'pageBot')
+                ).toBeUndefined();
+            });
+        });
+
+        describe('open_custom_portal', () => {
+            it('should add a global variable for the bot included in a open portal action', () => {
+                runtime.stateUpdated(
+                    stateUpdatedEvent({
+                        test1: createBot('test1', {
+                            abc: 'def',
+                        }),
+                    })
+                );
+                runtime.process([
+                    openCustomPortal('page', 'test1', 'myTag', {}),
+                ]);
+
+                expect((<any>globalThis).pageBot).toBe(
+                    runtime.context.state['test1']
+                );
+            });
+
+            it('should override previous variables', () => {
+                runtime.stateUpdated(
+                    stateUpdatedEvent({
+                        test1: createBot('test1', {
+                            abc: 'def',
+                        }),
+                        test2: createBot('test2', {
+                            abc: 'other',
+                        }),
+                    })
+                );
+                runtime.process([
+                    openCustomPortal('page', 'test1', 'myTag', {}),
+                ]);
+
+                expect((<any>globalThis).pageBot).toBe(
+                    runtime.context.state['test1']
+                );
+
+                runtime.process([
+                    openCustomPortal('page', 'test2', 'myTag', {}),
+                ]);
+
+                expect((<any>globalThis).pageBot).toBe(
+                    runtime.context.state['test2']
+                );
+            });
+
+            it('should remove the variable if given no bot to use for configuration', () => {
+                runtime.stateUpdated(
+                    stateUpdatedEvent({
+                        test1: createBot('test1', {
+                            abc: 'def',
+                        }),
+                    })
+                );
+                runtime.process([
+                    openCustomPortal('page', 'test1', 'myTag', {}),
+                ]);
+
+                expect((<any>globalThis).pageBot).toBe(
+                    runtime.context.state['test1']
+                );
+
+                runtime.process([openCustomPortal('page', null, 'myTag', {})]);
+
+                expect((<any>globalThis).pageBot).toBeUndefined();
+            });
+
+            it('should remove the global variables that were created by the runtime', () => {
+                runtime.stateUpdated(
+                    stateUpdatedEvent({
+                        test1: createBot('test1', {
+                            abc: 'def',
+                        }),
+                    })
+                );
+                runtime.process([
+                    openCustomPortal('page', 'test1', 'myTag', {}),
+                ]);
+
+                expect((<any>globalThis).pageBot).toBe(
+                    runtime.context.state['test1']
+                );
+
+                runtime.unsubscribe();
+
+                expect(
+                    Object.getOwnPropertyDescriptor(globalThis, 'pageBot')
+                ).toBeUndefined();
+            });
+
+            it('should emit open portal actions', async () => {
+                let actions = [] as BotAction[];
+                runtime.onActions.subscribe((a) => actions.push(...a));
+
+                runtime.process([
+                    openCustomPortal('page', 'test1', 'myTag', {}),
+                ]);
+
+                await waitAsync();
+
+                expect(actions).toEqual([
+                    openCustomPortal('page', 'test1', 'myTag', {}),
+                ]);
+            });
+        });
     });
 
     describe('execute()', () => {
         it('should compile and run the given script', async () => {
-            runtime.execute('player.toast("hello")');
+            runtime.execute('os.toast("hello")');
 
             await waitAsync();
 
@@ -3376,7 +3544,7 @@ describe('AuxRuntime', () => {
         });
 
         it('should emit an error if the script has a syntax error', async () => {
-            runtime.execute('player.toast(');
+            runtime.execute('os.toast(');
 
             await waitAsync();
 
@@ -3384,7 +3552,7 @@ describe('AuxRuntime', () => {
                 [
                     {
                         error: expect.any(SyntaxError),
-                        script: 'player.toast(',
+                        script: 'os.toast(',
                         bot: null,
                         tag: null,
                     },
@@ -3393,7 +3561,7 @@ describe('AuxRuntime', () => {
         });
 
         it('should return the compiler error if the script was unable to be compiled', async () => {
-            const result = runtime.execute('player.toast(');
+            const result = runtime.execute('os.toast(');
 
             await waitAsync();
 
@@ -3403,7 +3571,7 @@ describe('AuxRuntime', () => {
                 errors: [
                     {
                         error: expect.any(SyntaxError),
-                        script: 'player.toast(',
+                        script: 'os.toast(',
                         bot: null,
                         tag: null,
                     },
@@ -3417,10 +3585,10 @@ describe('AuxRuntime', () => {
             runtime.stateUpdated(
                 stateUpdatedEvent({
                     test1: createBot('test1', {
-                        hello: '@player.toast("hi1")',
+                        hello: '@os.toast("hi1")',
                     }),
                     test2: createBot('test2', {
-                        hello: '@player.toast("hi2")',
+                        hello: '@os.toast("hi2")',
                     }),
                     test3: createBot('test3', {}),
                 })
@@ -3436,10 +3604,10 @@ describe('AuxRuntime', () => {
             runtime.stateUpdated(
                 stateUpdatedEvent({
                     test1: createBot('test1', {
-                        hello: '@player.toast("hi1")',
+                        hello: '@os.toast("hi1")',
                     }),
                     test2: createBot('test2', {
-                        hello: '@player.toast("hi2")',
+                        hello: '@os.toast("hi2")',
                     }),
                     test3: createBot('test3', {}),
                 })
@@ -3455,7 +3623,7 @@ describe('AuxRuntime', () => {
             runtime.stateUpdated(
                 stateUpdatedEvent({
                     test1: createBot('test1', {
-                        hello: '@player.toast(that.toJSON())',
+                        hello: '@os.toast(that.toJSON())',
                     }),
                     test2: createBot('test2', {
                         abc: 'def',
@@ -3498,7 +3666,7 @@ describe('AuxRuntime', () => {
             runtime.stateUpdated(
                 stateUpdatedEvent({
                     test1: createBot('test1', {
-                        hello: '@player.toast(that)',
+                        hello: '@os.toast(that)',
                     }),
                 })
             );
@@ -3524,7 +3692,7 @@ describe('AuxRuntime', () => {
             runtime.stateUpdated(
                 stateUpdatedEvent({
                     test1: createBot('test1', {
-                        hello: '@player.toast(that)',
+                        hello: '@os.toast(that)',
                     }),
                 })
             );
@@ -3548,7 +3716,7 @@ describe('AuxRuntime', () => {
             runtime.stateUpdated(
                 stateUpdatedEvent({
                     test1: createBot('test1', {
-                        hello: '@player.toast(that)',
+                        hello: '@os.toast(that)',
                     }),
                 })
             );
@@ -3572,7 +3740,7 @@ describe('AuxRuntime', () => {
             runtime.stateUpdated(
                 stateUpdatedEvent({
                     test1: createBot('test1', {
-                        hello: '@player.toast(that)',
+                        hello: '@os.toast(that)',
                     }),
                 })
             );
@@ -3596,7 +3764,7 @@ describe('AuxRuntime', () => {
             runtime.stateUpdated(
                 stateUpdatedEvent({
                     test1: createBot('test1', {
-                        hello: '@player.toast(that)',
+                        hello: '@os.toast(that)',
                     }),
                 })
             );
@@ -3620,7 +3788,7 @@ describe('AuxRuntime', () => {
             runtime.stateUpdated(
                 stateUpdatedEvent({
                     test1: createBot('test1', {
-                        hello: '@player.toast(that)',
+                        hello: '@os.toast(that)',
                     }),
                 })
             );
@@ -3653,8 +3821,7 @@ describe('AuxRuntime', () => {
                 runtime.stateUpdated(
                     stateUpdatedEvent({
                         test1: createBot('test1', {
-                            hello:
-                                '@setInterval(() => player.toast("abc"), 100)',
+                            hello: '@setInterval(() => os.toast("abc"), 100)',
                         }),
                     })
                 );
@@ -3673,8 +3840,7 @@ describe('AuxRuntime', () => {
                 runtime.stateUpdated(
                     stateUpdatedEvent({
                         test1: createBot('test1', {
-                            hello:
-                                '@setTimeout(() => player.toast("abc"), 100)',
+                            hello: '@setTimeout(() => os.toast("abc"), 100)',
                         }),
                     })
                 );
@@ -3719,7 +3885,7 @@ describe('AuxRuntime', () => {
                 stateUpdatedEvent({
                     test1: createBot('test1', {
                         hello:
-                            '@Promise.resolve(0).then(() => player.toast("abc")).then(() => player.toast("abc2"))',
+                            '@Promise.resolve(0).then(() => os.toast("abc")).then(() => os.toast("abc2"))',
                     }),
                 })
             );
@@ -3735,7 +3901,7 @@ describe('AuxRuntime', () => {
                 stateUpdatedEvent({
                     test1: createBot('test1', {
                         hello: `@await Promise.resolve(0);
-                        player.toast("abc");`,
+                        os.toast("abc");`,
                     }),
                 })
             );
@@ -3751,7 +3917,7 @@ describe('AuxRuntime', () => {
                 stateUpdatedEvent({
                     test1: createBot('test1', {
                         hello: `@await Promise.resolve(0);
-                        player.toast("abc");
+                        os.toast("abc");
 
                         // Never gets resolved but that is fine because we
                         // want to ensure that the toast happens
@@ -3890,7 +4056,7 @@ describe('AuxRuntime', () => {
                 runtime.stateUpdated(
                     stateUpdatedEvent({
                         test1: createBot('test1', {
-                            create: `@create({ "shout": "@player.toast('abc')" })`,
+                            create: `@create({ "shout": "@os.toast('abc')" })`,
                         }),
                     })
                 );
@@ -3904,7 +4070,7 @@ describe('AuxRuntime', () => {
                         botAdded(
                             createBot('uuid', {
                                 creator: 'test1',
-                                shout: "@player.toast('abc')",
+                                shout: "@os.toast('abc')",
                             })
                         ),
                         toast('abc'),
@@ -3917,7 +4083,7 @@ describe('AuxRuntime', () => {
                 runtime.stateUpdated(
                     stateUpdatedEvent({
                         test1: createBot('test1', {
-                            create: `@create({ abc: "def", "shout": "@player.toast('abc')" })`,
+                            create: `@create({ abc: "def", "shout": "@os.toast('abc')" })`,
                         }),
                     })
                 );
@@ -3937,7 +4103,7 @@ describe('AuxRuntime', () => {
                         state: {
                             uuid: createPrecalculatedBot('uuid', {
                                 creator: 'test1',
-                                shout: "@player.toast('abc')",
+                                shout: "@os.toast('abc')",
                                 abc: 'def',
                             }),
                         },
@@ -4045,7 +4211,7 @@ describe('AuxRuntime', () => {
                     stateUpdatedEvent({
                         test1: createBot('test1', {
                             create:
-                                '@create({ abc: "@player.toast(`Hi`);" }); shout("abc");',
+                                '@create({ abc: "@os.toast(`Hi`);" }); shout("abc");',
                         }),
                     })
                 );
@@ -4058,7 +4224,7 @@ describe('AuxRuntime', () => {
                         botAdded(
                             createBot('uuid', {
                                 creator: 'test1',
-                                abc: '@player.toast(`Hi`);',
+                                abc: '@os.toast(`Hi`);',
                             })
                         ),
                         toast('Hi'),
@@ -4072,7 +4238,7 @@ describe('AuxRuntime', () => {
                     stateUpdatedEvent({
                         test1: createBot('test1', {
                             create:
-                                '@let created = create({ abc: "@player.toast(`Hi`);" }); whisper(created, "abc");',
+                                '@let created = create({ abc: "@os.toast(`Hi`);" }); whisper(created, "abc");',
                         }),
                     })
                 );
@@ -4085,7 +4251,7 @@ describe('AuxRuntime', () => {
                         botAdded(
                             createBot('uuid', {
                                 creator: 'test1',
-                                abc: '@player.toast(`Hi`);',
+                                abc: '@os.toast(`Hi`);',
                             })
                         ),
                         toast('Hi'),
@@ -4128,7 +4294,7 @@ describe('AuxRuntime', () => {
                     stateUpdatedEvent({
                         test1: createBot('test1', {
                             create:
-                                '@await testPromise; let created = create({ abc: "@player.toast(`Hi`);" }); whisper(created, "abc");',
+                                '@await testPromise; let created = create({ abc: "@os.toast(`Hi`);" }); whisper(created, "abc");',
                         }),
                     })
                 );
@@ -4147,7 +4313,7 @@ describe('AuxRuntime', () => {
                         botAdded(
                             createBot('uuid', {
                                 creator: 'test1',
-                                abc: '@player.toast(`Hi`);',
+                                abc: '@os.toast(`Hi`);',
                             })
                         ),
                         toast('Hi'),
@@ -4293,7 +4459,7 @@ describe('AuxRuntime', () => {
                     stateUpdatedEvent({
                         test1: createBot('test1', {
                             delete: '@destroy(this)',
-                            hello: '@player.toast("hi")',
+                            hello: '@os.toast("hi")',
                         }),
                     })
                 );
@@ -5106,7 +5272,7 @@ describe('AuxRuntime', () => {
                             tags.value = 123;
                             // value is not updated to 123 because
                             // the update is delayed
-                            player.toast(tags.value);
+                            os.toast(tags.value);
                         `,
                             },
                             <any>'delayed'
@@ -5635,18 +5801,18 @@ describe('AuxRuntime', () => {
             runtime.stateUpdated(
                 stateUpdatedEvent({
                     test: createBot('test', {
-                        script: '@player.toast("abc")',
+                        script: '@os.toast("abc")',
                     }),
                     test2: {
                         id: 'test2',
                         tags: {
-                            script: '@player.toast("def")',
+                            script: '@os.toast("def")',
                         },
                         signatures: {
                             [tagValueHash(
                                 'test2',
                                 'script',
-                                '@player.toast("def")'
+                                '@os.toast("def")'
                             )]: 'script',
                         },
                     },
@@ -5661,12 +5827,12 @@ describe('AuxRuntime', () => {
             runtime.stateUpdated(
                 stateUpdatedEvent({
                     test: createBot('test', {
-                        script: '@player.toast("abc")',
+                        script: '@os.toast("abc")',
                     }),
                 })
             );
 
-            const hash = tagValueHash('test', 'script', '@player.toast("abc")');
+            const hash = tagValueHash('test', 'script', '@os.toast("abc")');
 
             runtime.stateUpdated(
                 stateUpdatedEvent({
@@ -5688,20 +5854,20 @@ describe('AuxRuntime', () => {
                     test: {
                         id: 'test',
                         tags: {
-                            script: '@player.toast("def")',
+                            script: '@os.toast("def")',
                         },
                         signatures: {
                             [tagValueHash(
                                 'test',
                                 'script',
-                                '@player.toast("def")'
+                                '@os.toast("def")'
                             )]: 'script',
                         },
                     },
                 })
             );
 
-            const hash = tagValueHash('test', 'script', '@player.toast("def")');
+            const hash = tagValueHash('test', 'script', '@os.toast("def")');
 
             runtime.stateUpdated(
                 stateUpdatedEvent({
@@ -5723,7 +5889,7 @@ describe('AuxRuntime', () => {
                     test: createBot(
                         'test',
                         {
-                            script: '@player.toast("abc")',
+                            script: '@os.toast("abc")',
                         },
                         'tempLocal'
                     ),
@@ -5740,7 +5906,7 @@ describe('AuxRuntime', () => {
                     test: createBot(
                         'test',
                         {
-                            script: '@player.toast("abc")',
+                            script: '@os.toast("abc")',
                         },
                         'local'
                     ),
@@ -6189,7 +6355,7 @@ describe('original action tests', () => {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: `@player.toast("abc")\nthrow new Error("abc")`,
+                        test: `@os.toast("abc")\nthrow new Error("abc")`,
                     },
                 },
             };
@@ -6375,94 +6541,30 @@ describe('original action tests', () => {
         });
     });
 
-    describe('config', () => {
-        it('should pass in a config variable which equals getBot("id", tags.configBot)', () => {
+    describe('configBot', () => {
+        it('should get the current users bot', () => {
             const state: BotsState = {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        configBot: 'thatBot',
-                        test: '@setTag(this, "configId", config.id)',
+                        test: '@setTag(configBot, "#name", "Test")',
                     },
                 },
-                thatBot: {
-                    id: 'thatBot',
-                    tags: {
-                        name: 'Joe',
-                    },
+                userBot: {
+                    id: 'userBot',
+                    tags: {},
                 },
             };
 
             // specify the UUID to use next
             uuidMock.mockReturnValue('uuid-0');
-            const botAction = action('test', ['thisBot']);
+            const botAction = action('test', ['thisBot', 'userBot'], 'userBot');
             const result = calculateActionResults(state, botAction);
 
             expect(result.actions).toEqual([
-                botUpdated('thisBot', {
+                botUpdated('userBot', {
                     tags: {
-                        configId: 'thatBot',
-                    },
-                }),
-            ]);
-        });
-
-        it('the config variable should be null if configBot is not set', () => {
-            const state: BotsState = {
-                thisBot: {
-                    id: 'thisBot',
-                    tags: {
-                        test: '@setTag(this, "hasConfig", config !== null)',
-                    },
-                },
-                thatBot: {
-                    id: 'thatBot',
-                    tags: {
-                        name: 'Joe',
-                    },
-                },
-            };
-
-            // specify the UUID to use next
-            uuidMock.mockReturnValue('uuid-0');
-            const botAction = action('test', ['thisBot']);
-            const result = calculateActionResults(state, botAction);
-
-            expect(result.actions).toEqual([
-                botUpdated('thisBot', {
-                    tags: {
-                        hasConfig: false,
-                    },
-                }),
-            ]);
-        });
-
-        it('the config variable should be null if configBot is referencing a missing bot', () => {
-            const state: BotsState = {
-                thisBot: {
-                    id: 'thisBot',
-                    tags: {
-                        configBot: 'none',
-                        test: '@setTag(this, "hasConfig", config !== null)',
-                    },
-                },
-                thatBot: {
-                    id: 'thatBot',
-                    tags: {
-                        name: 'Joe',
-                    },
-                },
-            };
-
-            // specify the UUID to use next
-            uuidMock.mockReturnValue('uuid-0');
-            const botAction = action('test', ['thisBot']);
-            const result = calculateActionResults(state, botAction);
-
-            expect(result.actions).toEqual([
-                botUpdated('thisBot', {
-                    tags: {
-                        hasConfig: false,
+                        name: 'Test',
                     },
                 }),
             ]);
@@ -6489,77 +6591,6 @@ describe('original action tests', () => {
                 botUpdated('thisBot', {
                     tags: {
                         runningTag: 'test',
-                    },
-                }),
-            ]);
-        });
-    });
-
-    describe('configTag', () => {
-        it('should define a configTag variable which equals config.tags[tagName]', () => {
-            const state: BotsState = {
-                thisBot: {
-                    id: 'thisBot',
-                    tags: {
-                        configBot: 'thatBot',
-                        test: '@setTag(this, "parentScript", configTag)',
-                    },
-                },
-                thatBot: {
-                    id: 'thatBot',
-                    tags: {
-                        test: 'player.toast("hello")',
-                        name: 'Joe',
-                    },
-                },
-            };
-
-            // specify the UUID to use next
-            uuidMock.mockReturnValue('uuid-0');
-            const botAction = action('test', ['thisBot']);
-            const result = calculateActionResults(state, botAction);
-
-            expect(result.actions).toEqual([
-                botUpdated('thisBot', {
-                    tags: {
-                        parentScript: 'player.toast("hello")',
-                    },
-                }),
-            ]);
-        });
-
-        it('should contain the value that the config bot contained when the script started', () => {
-            const state: BotsState = {
-                thisBot: {
-                    id: 'thisBot',
-                    tags: {
-                        configBot: 'thatBot',
-                        test: `@config.tags.test = "abc"; setTag(this, "parentScript", configTag)`,
-                    },
-                },
-                thatBot: {
-                    id: 'thatBot',
-                    tags: {
-                        test: 'player.toast("hello")',
-                        name: 'Joe',
-                    },
-                },
-            };
-
-            // specify the UUID to use next
-            uuidMock.mockReturnValue('uuid-0');
-            const botAction = action('test', ['thisBot']);
-            const result = calculateActionResults(state, botAction);
-
-            expect(result.actions).toEqual([
-                botUpdated('thatBot', {
-                    tags: {
-                        test: 'abc',
-                    },
-                }),
-                botUpdated('thisBot', {
-                    tags: {
-                        parentScript: 'player.toast("hello")',
                     },
                 }),
             ]);
@@ -6813,7 +6844,7 @@ describe('original action tests', () => {
                 id: 'thisBot',
                 tags: {
                     abcdef: '@shout("sayHello")',
-                    sayHello: '@setTag(this, "#userId", player.getBot().id)',
+                    sayHello: '@setTag(this, "#userId", configBot.id)',
                 },
             },
         };
@@ -6870,7 +6901,7 @@ describe('original action tests', () => {
             bBot: {
                 id: 'bBot',
                 tags: {
-                    test: `@player.toast("hello")`,
+                    test: `@os.toast("hello")`,
                 },
             },
         };
@@ -6897,7 +6928,7 @@ describe('original action tests', () => {
             bBot: {
                 id: 'bBot',
                 tags: {
-                    test: `@player.toast("hello")`,
+                    test: `@os.toast("hello")`,
                 },
             },
         };
@@ -6920,7 +6951,7 @@ describe('original action tests', () => {
             thisBot: {
                 id: 'thisBot',
                 tags: {
-                    test: '@player.toast("test"); // this is a test',
+                    test: '@os.toast("test"); // this is a test',
                 },
             },
         };
@@ -6944,8 +6975,7 @@ describe('original action tests', () => {
             thisBot: {
                 id: 'thisBot',
                 tags: {
-                    test:
-                        '@player.toast("test"); // comment 1\n// this is a test',
+                    test: '@os.toast("test"); // comment 1\n// this is a test',
                 },
             },
         };
@@ -6957,8 +6987,6 @@ describe('original action tests', () => {
 
         expect(events.actions).toEqual([toast('test')]);
     });
-
-    // describe('')
 
     describe('onAnyListen()', () => {
         it('should send a onAnyListen() for actions', () => {
@@ -7132,7 +7160,7 @@ describe('original action tests', () => {
                 bot1: {
                     id: 'bot1',
                     tags: {
-                        onAnyListen: `@player.toast('Hi!');`,
+                        onAnyListen: `@os.toast('Hi!');`,
                     },
                 },
                 bot2: {
@@ -7677,7 +7705,7 @@ describe('original action tests', () => {
                     tags: {
                         _position: { x: 0, y: 0, z: 0 },
                         _workspace: 'abc',
-                        abcdef: `@action.perform(player.toast('abc'))`,
+                        abcdef: `@action.perform(os.toast('abc'))`,
                     },
                 },
             };
@@ -7698,7 +7726,7 @@ describe('original action tests', () => {
                         _position: { x: 0, y: 0, z: 0 },
                         _workspace: 'abc',
                         abcdef: `@
-                            const toast = player.toast('abc');
+                            const toast = os.toast('abc');
                             action.reject(toast);
                             action.perform(toast);
                         `,
@@ -8680,8 +8708,7 @@ describe('original action tests', () => {
                     id: 'thisBot',
                     tags: {
                         abc: true,
-                        test:
-                            '@destroy(this); player.toast(getBot("abc", true));',
+                        test: '@destroy(this); os.toast(getBot("abc", true));',
                     },
                 },
             };
@@ -8884,14 +8911,14 @@ describe('original action tests', () => {
         });
     });
 
-    describe('player.getDimensionalDepth()', () => {
+    describe('os.getDimensionalDepth()', () => {
         it('should return 0 when the bot is in the given dimension', () => {
             const state: BotsState = {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
                         test:
-                            '@setTag(this, "depth", player.getDimensionalDepth("dimension"))',
+                            '@setTag(this, "depth", os.getDimensionalDepth("dimension"))',
                     },
                 },
                 userBot: {
@@ -8926,7 +8953,7 @@ describe('original action tests', () => {
                         id: 'thisBot',
                         tags: {
                             test:
-                                '@setTag(this, "depth", player.getDimensionalDepth("dimension"))',
+                                '@setTag(this, "depth", os.getDimensionalDepth("dimension"))',
                         },
                     },
                     userBot: {
@@ -8962,7 +8989,7 @@ describe('original action tests', () => {
                     id: 'thisBot',
                     tags: {
                         test:
-                            '@setTag(this, "depth", player.getDimensionalDepth("dimension"))',
+                            '@setTag(this, "depth", os.getDimensionalDepth("dimension"))',
                     },
                 },
                 userBot: {
@@ -8986,44 +9013,16 @@ describe('original action tests', () => {
         });
     });
 
-    describe('player.getBot()', () => {
-        it('should get the current users bot', () => {
-            const state: BotsState = {
-                thisBot: {
-                    id: 'thisBot',
-                    tags: {
-                        test: '@setTag(player.getBot(), "#name", "Test")',
-                    },
-                },
-                userBot: {
-                    id: 'userBot',
-                    tags: {},
-                },
-            };
+    describe('player.getBot()', () => {});
 
-            // specify the UUID to use next
-            uuidMock.mockReturnValue('uuid-0');
-            const botAction = action('test', ['thisBot', 'userBot'], 'userBot');
-            const result = calculateActionResults(state, botAction);
-
-            expect(result.actions).toEqual([
-                botUpdated('userBot', {
-                    tags: {
-                        name: 'Test',
-                    },
-                }),
-            ]);
-        });
-    });
-
-    describe('player.replaceDragBot()', () => {
+    describe('os.replaceDragBot()', () => {
         it('should send a replace_drag_bot event', () => {
             const state: BotsState = {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
                         abc: true,
-                        test: '@player.replaceDragBot(this)',
+                        test: '@os.replaceDragBot(this)',
                     },
                 },
             };
@@ -9041,7 +9040,7 @@ describe('original action tests', () => {
                     id: 'thisBot',
                     tags: {
                         abc: true,
-                        test: '@player.replaceDragBot(this)',
+                        test: '@os.replaceDragBot(this)',
                     },
                 },
             };
@@ -9302,7 +9301,7 @@ describe('original action tests', () => {
                     id: 'thisBot',
                     tags: {
                         test:
-                            '@setTag(this, "#dimension", player.getMenuDimension())',
+                            '@setTag(this, "#dimension", os.getMenuDimension())',
                     },
                 },
                 userBot: {
@@ -9328,13 +9327,13 @@ describe('original action tests', () => {
         });
     });
 
-    describe('player.toast()', () => {
+    describe('os.toast()', () => {
         it('should emit a ShowToastAction', () => {
             const state: BotsState = {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@player.toast("hello, world!")',
+                        test: '@os.toast("hello, world!")',
                     },
                 },
             };
@@ -9348,13 +9347,13 @@ describe('original action tests', () => {
         });
     });
 
-    describe('player.showJoinCode()', () => {
+    describe('os.showJoinCode()', () => {
         it('should emit a ShowJoinCodeEvent', () => {
             const state: BotsState = {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@player.showJoinCode()',
+                        test: '@os.showJoinCode()',
                     },
                 },
             };
@@ -9372,7 +9371,7 @@ describe('original action tests', () => {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@player.showJoinCode("server", "dimension")',
+                        test: '@os.showJoinCode("server", "dimension")',
                     },
                 },
             };
@@ -9388,13 +9387,13 @@ describe('original action tests', () => {
         });
     });
 
-    describe('player.requestFullscreenMode()', () => {
+    describe('os.requestFullscreenMode()', () => {
         it('should issue a request_fullscreen action', () => {
             const state: BotsState = {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@player.requestFullscreenMode()',
+                        test: '@os.requestFullscreenMode()',
                     },
                 },
             };
@@ -9408,13 +9407,13 @@ describe('original action tests', () => {
         });
     });
 
-    describe('player.exitFullscreenMode()', () => {
+    describe('os.exitFullscreenMode()', () => {
         it('should issue a request_fullscreen action', () => {
             const state: BotsState = {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@player.exitFullscreenMode()',
+                        test: '@os.exitFullscreenMode()',
                     },
                 },
             };
@@ -9428,13 +9427,13 @@ describe('original action tests', () => {
         });
     });
 
-    describe('player.showHtml()', () => {
+    describe('os.showHtml()', () => {
         it('should issue a show_html action', () => {
             const state: BotsState = {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@player.showHtml("hello, world!")',
+                        test: '@os.showHtml("hello, world!")',
                     },
                 },
             };
@@ -9448,13 +9447,13 @@ describe('original action tests', () => {
         });
     });
 
-    describe('player.hideHtml()', () => {
+    describe('os.hideHtml()', () => {
         it('should issue a hide_html action', () => {
             const state: BotsState = {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@player.hideHtml()',
+                        test: '@os.hideHtml()',
                     },
                 },
             };
@@ -9468,13 +9467,13 @@ describe('original action tests', () => {
         });
     });
 
-    describe('player.setClipboard()', () => {
+    describe('os.setClipboard()', () => {
         it('should emit a SetClipboardEvent', () => {
             const state: BotsState = {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@player.setClipboard("test")',
+                        test: '@os.setClipboard("test")',
                     },
                 },
             };
@@ -9488,13 +9487,13 @@ describe('original action tests', () => {
         });
     });
 
-    describe('player.tweenTo()', () => {
+    describe('os.tweenTo()', () => {
         it('should emit a TweenToAction', () => {
             const state: BotsState = {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@player.tweenTo("test")',
+                        test: '@os.tweenTo("test")',
                     },
                 },
             };
@@ -9512,7 +9511,7 @@ describe('original action tests', () => {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@player.tweenTo(this)',
+                        test: '@os.tweenTo(this)',
                     },
                 },
             };
@@ -9531,7 +9530,7 @@ describe('original action tests', () => {
                     id: 'thisBot',
                     tags: {
                         test:
-                            '@player.tweenTo("test", undefined, undefined, undefined, 10)',
+                            '@os.tweenTo("test", undefined, undefined, undefined, 10)',
                     },
                 },
             };
@@ -9547,13 +9546,13 @@ describe('original action tests', () => {
         });
     });
 
-    describe('player.moveTo()', () => {
+    describe('os.moveTo()', () => {
         it('should emit a TweenToAction with the duration set to 0', () => {
             const state: BotsState = {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@player.moveTo("test")',
+                        test: '@os.moveTo("test")',
                     },
                 },
             };
@@ -9575,13 +9574,13 @@ describe('original action tests', () => {
         });
     });
 
-    describe('player.showChat()', () => {
+    describe('os.showChat()', () => {
         it('should emit a ShowChatBarAction', () => {
             const state: BotsState = {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@player.showChat()',
+                        test: '@os.showChat()',
                     },
                 },
             };
@@ -9599,7 +9598,7 @@ describe('original action tests', () => {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@player.showChat("test")',
+                        test: '@os.showChat("test")',
                     },
                 },
             };
@@ -9621,7 +9620,7 @@ describe('original action tests', () => {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: `@player.showChat({
+                        test: `@os.showChat({
                             placeholder: "abc",
                             prefill: "def"
                         })`,
@@ -9643,13 +9642,13 @@ describe('original action tests', () => {
         });
     });
 
-    describe('player.hideChat()', () => {
+    describe('os.hideChat()', () => {
         it('should emit a ShowChatBarAction', () => {
             const state: BotsState = {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@player.hideChat()',
+                        test: '@os.hideChat()',
                     },
                 },
             };
@@ -9663,13 +9662,13 @@ describe('original action tests', () => {
         });
     });
 
-    describe('player.run()', () => {
+    describe('os.run()', () => {
         it('should emit a RunScriptAction', () => {
             const state: BotsState = {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@player.run("abc")',
+                        test: '@os.run("abc")',
                     },
                 },
             };
@@ -9683,13 +9682,13 @@ describe('original action tests', () => {
         });
     });
 
-    describe('player.version()', () => {
+    describe('os.version()', () => {
         it('should return an object with version information', () => {
             const state: BotsState = {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@tags.version = player.version()',
+                        test: '@tags.version = os.version()',
                     },
                 },
             };
@@ -9721,13 +9720,13 @@ describe('original action tests', () => {
         });
     });
 
-    describe('player.device()', () => {
+    describe('os.device()', () => {
         it('should return an object with device information', () => {
             const state: BotsState = {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@tags.device = player.device()',
+                        test: '@tags.device = os.device()',
                     },
                 },
             };
@@ -9757,7 +9756,7 @@ describe('original action tests', () => {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@tags.device = player.device()',
+                        test: '@tags.device = os.device()',
                     },
                 },
             };
@@ -9780,13 +9779,13 @@ describe('original action tests', () => {
         });
     });
 
-    describe('player.enableAR()', () => {
+    describe('os.enableAR()', () => {
         it('should issue an EnableARAction', () => {
             const state: BotsState = {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@player.enableAR()',
+                        test: '@os.enableAR()',
                     },
                 },
             };
@@ -9800,13 +9799,13 @@ describe('original action tests', () => {
         });
     });
 
-    describe('player.disableAR()', () => {
+    describe('os.disableAR()', () => {
         it('should issue an EnableVRAction', () => {
             const state: BotsState = {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@player.disableAR()',
+                        test: '@os.disableAR()',
                     },
                 },
             };
@@ -9820,13 +9819,13 @@ describe('original action tests', () => {
         });
     });
 
-    describe('player.enableVR()', () => {
+    describe('os.enableVR()', () => {
         it('should issue an EnableVRAction', () => {
             const state: BotsState = {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@player.enableVR()',
+                        test: '@os.enableVR()',
                     },
                 },
             };
@@ -9840,13 +9839,13 @@ describe('original action tests', () => {
         });
     });
 
-    describe('player.disableVR()', () => {
+    describe('os.disableVR()', () => {
         it('should issue an EnableVRAction', () => {
             const state: BotsState = {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@player.disableVR()',
+                        test: '@os.disableVR()',
                     },
                 },
             };
@@ -9860,14 +9859,14 @@ describe('original action tests', () => {
         });
     });
 
-    describe('player.downloadBots()', () => {
+    describe('os.downloadBots()', () => {
         it('should emit a DownloadAction with the given bots formatted as JSON', () => {
             const state: BotsState = {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
                         test:
-                            '@player.downloadBots(getBots(inDimension("abc")), "test")',
+                            '@os.downloadBots(getBots(inDimension("abc")), "test")',
                     },
                 },
                 funBot: {
@@ -9912,7 +9911,7 @@ describe('original action tests', () => {
                     id: 'thisBot',
                     tags: {
                         test:
-                            '@player.downloadBots(getBots(inDimension("abc")), "test.aux")',
+                            '@os.downloadBots(getBots(inDimension("abc")), "test.aux")',
                     },
                 },
                 funBot: {
@@ -9952,13 +9951,13 @@ describe('original action tests', () => {
         });
     });
 
-    describe('player.showUploadAuxFile()', () => {
+    describe('os.showUploadAuxFile()', () => {
         it('should emit a showUploadAuxFileAction', () => {
             const state: BotsState = {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@player.showUploadAuxFile()',
+                        test: '@os.showUploadAuxFile()',
                     },
                 },
             };
@@ -9972,13 +9971,13 @@ describe('original action tests', () => {
         });
     });
 
-    describe('player.downloadServer()', () => {
+    describe('os.downloadServer()', () => {
         it('should emit a DownloadAction with the current state and server name', () => {
             const state: BotsState = {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@player.downloadServer()',
+                        test: '@os.downloadServer()',
                     },
                 },
                 userBot: {
@@ -10011,7 +10010,7 @@ describe('original action tests', () => {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@player.downloadServer()',
+                        test: '@os.downloadServer()',
                     },
                 },
                 thatBot: {
@@ -10071,7 +10070,7 @@ describe('original action tests', () => {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@player.openQRCodeScanner()',
+                        test: '@os.openQRCodeScanner()',
                     },
                 },
             };
@@ -10089,7 +10088,7 @@ describe('original action tests', () => {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@player.openQRCodeScanner("front")',
+                        test: '@os.openQRCodeScanner("front")',
                     },
                 },
             };
@@ -10109,7 +10108,7 @@ describe('original action tests', () => {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@player.closeQRCodeScanner()',
+                        test: '@os.closeQRCodeScanner()',
                     },
                 },
             };
@@ -10129,7 +10128,7 @@ describe('original action tests', () => {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@player.showQRCode("hello")',
+                        test: '@os.showQRCode("hello")',
                     },
                 },
             };
@@ -10149,7 +10148,7 @@ describe('original action tests', () => {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@player.hideQRCode()',
+                        test: '@os.hideQRCode()',
                     },
                 },
             };
@@ -10169,7 +10168,7 @@ describe('original action tests', () => {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@player.openBarcodeScanner()',
+                        test: '@os.openBarcodeScanner()',
                     },
                 },
             };
@@ -10187,7 +10186,7 @@ describe('original action tests', () => {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@player.openBarcodeScanner("front")',
+                        test: '@os.openBarcodeScanner("front")',
                     },
                 },
             };
@@ -10207,7 +10206,7 @@ describe('original action tests', () => {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@player.closeBarcodeScanner()',
+                        test: '@os.closeBarcodeScanner()',
                     },
                 },
             };
@@ -10227,7 +10226,7 @@ describe('original action tests', () => {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@player.showBarcode("hello")',
+                        test: '@os.showBarcode("hello")',
                     },
                 },
             };
@@ -10245,7 +10244,7 @@ describe('original action tests', () => {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@player.showBarcode("hello", "format")',
+                        test: '@os.showBarcode("hello", "format")',
                     },
                 },
             };
@@ -10267,7 +10266,7 @@ describe('original action tests', () => {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@player.hideBarcode()',
+                        test: '@os.hideBarcode()',
                     },
                 },
             };
@@ -10287,7 +10286,7 @@ describe('original action tests', () => {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@player.loadServer("abc")',
+                        test: '@os.loadServer("abc")',
                     },
                 },
             };
@@ -10307,7 +10306,7 @@ describe('original action tests', () => {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@player.unloadServer("abc")',
+                        test: '@os.unloadServer("abc")',
                     },
                 },
             };
@@ -10327,7 +10326,7 @@ describe('original action tests', () => {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@player.importAUX("abc")',
+                        test: '@os.importAUX("abc")',
                     },
                 },
             };
@@ -10357,7 +10356,7 @@ describe('original action tests', () => {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: `@player.importAUX('${json}')`,
+                        test: `@os.importAUX('${json}')`,
                     },
                 },
             };
@@ -10371,14 +10370,14 @@ describe('original action tests', () => {
         });
     });
 
-    describe('player.isInDimension()', () => {
+    describe('os.isInDimension()', () => {
         it('should return true when pagePortal equals the given value', () => {
             const state: BotsState = {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
                         test:
-                            '@setTag(this, "#inDimension", player.isInDimension("dimension"))',
+                            '@setTag(this, "#inDimension", os.isInDimension("dimension"))',
                     },
                 },
                 userBot: {
@@ -10409,7 +10408,7 @@ describe('original action tests', () => {
                     id: 'thisBot',
                     tags: {
                         test:
-                            '@setTag(this, "#inDimension", player.isInDimension("abc"))',
+                            '@setTag(this, "#inDimension", os.isInDimension("abc"))',
                     },
                 },
                 userBot: {
@@ -10440,7 +10439,7 @@ describe('original action tests', () => {
                     id: 'thisBot',
                     tags: {
                         test:
-                            '@setTag(this, "#inDimension", player.isInDimension("abc"))',
+                            '@setTag(this, "#inDimension", os.isInDimension("abc"))',
                     },
                 },
                 userBot: {
@@ -10464,14 +10463,14 @@ describe('original action tests', () => {
         });
     });
 
-    describe('player.getCurrentDimension()', () => {
+    describe('os.getCurrentDimension()', () => {
         it('should return pagePortal', () => {
             const state: BotsState = {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
                         test:
-                            '@setTag(this, "#dimension", player.getCurrentDimension())',
+                            '@setTag(this, "#dimension", os.getCurrentDimension())',
                     },
                 },
                 userBot: {
@@ -10502,7 +10501,7 @@ describe('original action tests', () => {
                     id: 'thisBot',
                     tags: {
                         test:
-                            '@setTag(this, "#dimension", player.getCurrentDimension())',
+                            '@setTag(this, "#dimension", os.getCurrentDimension())',
                     },
                 },
                 userBot: {
@@ -10526,14 +10525,14 @@ describe('original action tests', () => {
         });
     });
 
-    describe('player.getCurrentServer()', () => {
+    describe('os.getCurrentServer()', () => {
         it('should return server', () => {
             const state: BotsState = {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
                         test:
-                            '@setTag(this, "#dimension", player.getCurrentServer())',
+                            '@setTag(this, "#dimension", os.getCurrentServer())',
                     },
                 },
                 userBot: {
@@ -10564,7 +10563,7 @@ describe('original action tests', () => {
                     id: 'thisBot',
                     tags: {
                         test:
-                            '@setTag(this, "#dimension", player.getCurrentServer())',
+                            '@setTag(this, "#dimension", os.getCurrentServer())',
                     },
                 },
                 userBot: {
@@ -10598,7 +10597,7 @@ describe('original action tests', () => {
                         id: 'thisBot',
                         tags: {
                             test:
-                                '@setTag(this, "#dimension", player.getCurrentServer())',
+                                '@setTag(this, "#dimension", os.getCurrentServer())',
                         },
                     },
                     userBot: {
@@ -10625,7 +10624,7 @@ describe('original action tests', () => {
         );
     });
 
-    describe('player.getPortalDimension()', () => {
+    describe('os.getPortalDimension()', () => {
         const cases = [
             ['page', 'pageDimension'],
             ['pagePortal', 'pageDimension'],
@@ -10646,7 +10645,7 @@ describe('original action tests', () => {
                     thisBot: {
                         id: 'thisBot',
                         tags: {
-                            test: `@setTag(this, "#fun", player.getPortalDimension("${portal}"))`,
+                            test: `@setTag(this, "#fun", os.getPortalDimension("${portal}"))`,
                         },
                     },
                     userBot: {
@@ -10678,13 +10677,13 @@ describe('original action tests', () => {
         );
     });
 
-    describe('player.showInputForTag()', () => {
+    describe('os.showInputForTag()', () => {
         it('should emit a ShowInputForTagAction', () => {
             const state: BotsState = {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@player.showInputForTag(this, "abc")',
+                        test: '@os.showInputForTag(this, "abc")',
                     },
                 },
             };
@@ -10702,7 +10701,7 @@ describe('original action tests', () => {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@player.showInputForTag("test", "abc")',
+                        test: '@os.showInputForTag("test", "abc")',
                     },
                 },
             };
@@ -10721,7 +10720,7 @@ describe('original action tests', () => {
                     id: 'thisBot',
                     tags: {
                         test:
-                            '@player.showInputForTag("test", "##abc"); player.showInputForTag("test", "#abc")',
+                            '@os.showInputForTag("test", "##abc"); os.showInputForTag("test", "#abc")',
                     },
                 },
             };
@@ -10743,7 +10742,7 @@ describe('original action tests', () => {
                     id: 'thisBot',
                     tags: {
                         test:
-                            '@player.showInputForTag("test", "abc", { backgroundColor: "red", foregroundColor: "green" })',
+                            '@os.showInputForTag("test", "abc", { backgroundColor: "red", foregroundColor: "green" })',
                     },
                 },
             };
@@ -10768,7 +10767,7 @@ describe('original action tests', () => {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@player.goToDimension("abc")',
+                        test: '@os.goToDimension("abc")',
                     },
                 },
             };
@@ -10786,7 +10785,7 @@ describe('original action tests', () => {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@player.goToDimension("sim", "abc")',
+                        test: '@os.goToDimension("sim", "abc")',
                     },
                 },
             };
@@ -10800,13 +10799,13 @@ describe('original action tests', () => {
         });
     });
 
-    describe('player.goToURL()', () => {
+    describe('os.goToURL()', () => {
         it('should issue a GoToURL event', () => {
             const state: BotsState = {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@player.goToURL("abc")',
+                        test: '@os.goToURL("abc")',
                     },
                 },
             };
@@ -10820,13 +10819,13 @@ describe('original action tests', () => {
         });
     });
 
-    describe('player.openURL()', () => {
+    describe('os.openURL()', () => {
         it('should issue a OpenURL event', () => {
             const state: BotsState = {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@player.openURL("abc")',
+                        test: '@os.openURL("abc")',
                     },
                 },
             };
@@ -10840,13 +10839,13 @@ describe('original action tests', () => {
         });
     });
 
-    describe('player.openDevConsole()', () => {
+    describe('os.openDevConsole()', () => {
         it('should issue a OpenConsole event', () => {
             const state: BotsState = {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: '@player.openDevConsole()',
+                        test: '@os.openDevConsole()',
                     },
                 },
             };
@@ -11314,13 +11313,13 @@ describe('original action tests', () => {
         });
     });
 
-    describe('player.checkout()', () => {
+    describe('os.checkout()', () => {
         it('should emit a start checkout event', () => {
             const state: BotsState = {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: `@player.checkout({
+                        test: `@os.checkout({
                             publishableKey: 'key',
                             productId: 'ID1',
                             title: 'Product 1',
@@ -11520,41 +11519,35 @@ describe('original action tests', () => {
 
     describe('remote()', () => {
         const cases = [
-            ['player.toast("My Message!")', toast('My Message!')] as const,
+            ['os.toast("My Message!")', toast('My Message!')] as const,
             [
-                'player.goToDimension("dimension")',
+                'os.goToDimension("dimension")',
                 goToDimension('dimension'),
             ] as const,
-            ['player.openURL("url")', openURL('url')] as const,
-            ['player.goToURL("url")', goToURL('url')] as const,
-            ['player.tweenTo("id")', tweenTo('id')] as const,
-            ['player.openURL("url")', openURL('url')] as const,
-            ['player.openQRCodeScanner()', openQRCodeScanner(true)] as const,
-            ['player.closeQRCodeScanner()', openQRCodeScanner(false)] as const,
-            ['player.openBarcodeScanner()', openBarcodeScanner(true)] as const,
+            ['os.openURL("url")', openURL('url')] as const,
+            ['os.goToURL("url")', goToURL('url')] as const,
+            ['os.tweenTo("id")', tweenTo('id')] as const,
+            ['os.openURL("url")', openURL('url')] as const,
+            ['os.openQRCodeScanner()', openQRCodeScanner(true)] as const,
+            ['os.closeQRCodeScanner()', openQRCodeScanner(false)] as const,
+            ['os.openBarcodeScanner()', openBarcodeScanner(true)] as const,
+            ['os.closeBarcodeScanner()', openBarcodeScanner(false)] as const,
+            ['os.showBarcode("code")', showBarcode(true, 'code')] as const,
+            ['os.hideBarcode()', showBarcode(false)] as const,
+            ['os.loadServer("channel")', loadSimulation('channel')] as const,
             [
-                'player.closeBarcodeScanner()',
-                openBarcodeScanner(false),
-            ] as const,
-            ['player.showBarcode("code")', showBarcode(true, 'code')] as const,
-            ['player.hideBarcode()', showBarcode(false)] as const,
-            [
-                'player.loadServer("channel")',
-                loadSimulation('channel'),
-            ] as const,
-            [
-                'player.unloadServer("channel")',
+                'os.unloadServer("channel")',
                 unloadSimulation('channel'),
             ] as const,
-            ['player.importAUX("aux")', importAUX('aux')] as const,
-            ['player.showQRCode("code")', showQRCode(true, 'code')] as const,
-            ['player.hideQRCode()', showQRCode(false)] as const,
+            ['os.importAUX("aux")', importAUX('aux')] as const,
+            ['os.showQRCode("code")', showQRCode(true, 'code')] as const,
+            ['os.hideQRCode()', showQRCode(false)] as const,
             [
-                'player.showInputForTag(this, "abc")',
+                'os.showInputForTag(this, "abc")',
                 showInputForTag('thisBot', 'abc'),
             ] as const,
             [
-                `player.checkout({
+                `os.checkout({
                 publishableKey: 'my_key',
                 productId: 'ID1',
                 title: 'Product 1',
@@ -11569,7 +11562,7 @@ describe('original action tests', () => {
                     processingServer: 'channel2',
                 }),
             ] as const,
-            ['player.openDevConsole()', openConsole()] as const,
+            ['os.openDevConsole()', openConsole()] as const,
         ];
 
         it.each(cases)('should wrap %s in a remote event', (script, event) => {
@@ -11595,7 +11588,7 @@ describe('original action tests', () => {
                 thisBot: {
                     id: 'thisBot',
                     tags: {
-                        test: `@remote(player.toast("Hi!"), {
+                        test: `@remote(os.toast("Hi!"), {
                             session: 's',
                             username: 'u',
                             device: 'd'
