@@ -44,6 +44,9 @@ import {
     DNA_TAG_PREFIX,
     UpdateBotAction,
     isRuntimeBot,
+    OpenCustomPortalAction,
+    createBot,
+    openCustomPortal,
 } from '../bots';
 import { Observable, Subject, Subscription, SubscriptionLike } from 'rxjs';
 import { AuxCompiler, AuxCompiledScript } from './AuxCompiler';
@@ -96,7 +99,7 @@ export class AuxRuntime
     private _errorBatch: ScriptError[] = [];
 
     private _userId: string;
-    private _sub: SubscriptionLike;
+    private _sub: Subscription;
     private _currentVersion: RuntimeStateVersion = {
         localSites: {},
         vector: {},
@@ -114,6 +117,7 @@ export class AuxRuntime
     private _forceSignedScripts: boolean;
     private _exemptSpaces: BotSpace[];
     private _batchPending: boolean = false;
+    private _portalBots: Map<string, string> = new Map();
 
     get forceSignedScripts() {
         return this._forceSignedScripts;
@@ -269,9 +273,50 @@ export class AuxRuntime
             this._globalContext.resolveTask(action.taskId, action.result, true);
         } else if (action.type === 'device_error') {
             this._globalContext.rejectTask(action.taskId, action.error, true);
+        } else if (action.type === 'open_custom_portal') {
+            this._registerPortalBot(action.portalId, action.botId);
+            this._actionBatch.push(action);
+        } else if (action.type === 'register_builtin_portal') {
+            if (!this._portalBots.has(action.portalId)) {
+                const newBot = this.context.createBot(
+                    createBot(undefined, undefined, 'tempLocal')
+                );
+                this._registerPortalBot(action.portalId, newBot.id);
+                this._actionBatch.push(
+                    openCustomPortal(action.portalId, newBot.id, null, {})
+                );
+            }
         } else {
             this._actionBatch.push(action);
         }
+    }
+
+    private _registerPortalBot(portalId: string, botId: string) {
+        this._portalBots.set(portalId, botId);
+        let anyGlobal: any = globalThis;
+        const variableName = `${portalId}Bot`;
+        if (hasValue(botId)) {
+            Object.defineProperty(anyGlobal, variableName, {
+                get: () => this.context.state[botId],
+                enumerable: false,
+                configurable: true,
+            });
+        } else {
+            delete anyGlobal[variableName];
+        }
+
+        this._sub.add(() => {
+            if (this._portalBots.get(portalId) === botId) {
+                const descriptor = Object.getOwnPropertyDescriptor(
+                    anyGlobal,
+                    variableName
+                );
+                if (descriptor) {
+                    delete anyGlobal[variableName];
+                    this._portalBots.delete(portalId);
+                }
+            }
+        });
     }
 
     private _rejectAction(
@@ -1226,14 +1271,10 @@ export class AuxRuntime
                 bot,
                 tag,
                 creator: null as RuntimeBot,
-                config: null as RuntimeBot,
             },
             before: (ctx) => {
                 ctx.creator = ctx.bot
                     ? this._getRuntimeBot(ctx.bot.script.tags.creator)
-                    : null;
-                ctx.config = ctx.bot
-                    ? this._getRuntimeBot(ctx.bot.script.tags.configBot)
                     : null;
             },
             onError: (err, ctx, meta) => {
@@ -1289,9 +1330,7 @@ export class AuxRuntime
                 raw: (ctx) => (ctx.bot ? ctx.bot.script.raw : null),
                 masks: (ctx) => (ctx.bot ? ctx.bot.script.masks : null),
                 creator: (ctx) => ctx.creator,
-                config: (ctx) => ctx.config,
-                configTag: (ctx) =>
-                    ctx.config ? ctx.config.tags[ctx.tag] : null,
+                configBot: () => this.context.playerBot,
             },
             arguments: [['that', 'data']],
         });
