@@ -2,16 +2,17 @@ import { createBot } from '@casual-simulation/aux-common';
 import { waitAsync } from '@casual-simulation/aux-common/test/TestHelpers';
 import { waitForSync } from '@casual-simulation/aux-vm';
 import { Simulation, SimulationManager } from '@casual-simulation/aux-vm';
+import { RemoteSimulation } from '@casual-simulation/aux-vm-client';
 import { nodeSimulationWithConfig } from '@casual-simulation/aux-vm-node';
 import { MenuPortal } from './MenuPortal';
 
 console.log = jest.fn();
 
 describe('MenuPortal', () => {
-    let simulationManager: SimulationManager<Simulation>;
+    let simulationManager: SimulationManager<RemoteSimulation>;
 
     beforeEach(() => {
-        simulationManager = new SimulationManager<Simulation>((id) =>
+        simulationManager = new SimulationManager<RemoteSimulation>((id) =>
             nodeSimulationWithConfig(
                 {
                     id: 'user',
