@@ -5,7 +5,7 @@
 'use strict';
 
 import type * as mode from './tsMode';
-import { typescriptVersion as tsversion } from './lib/typescriptServicesMetadata'; // do not import the whole typescriptServices here
+import { typescriptVersion as tsversion } from '@casual-simulation/aux-custom-portals/monaco/typescript/lib/typescriptServicesMetadata'; // do not import the whole typescriptServices here
 import {
     languages,
     Emitter,
