@@ -23,6 +23,7 @@ import RxjsOperatorsLibraryCode from '!raw-loader!@casual-simulation/aux-custom-
 import LodashLibraryCode from '!raw-loader!@casual-simulation/aux-custom-portals/dist/esbuild/lodash.js';
 import UuidLibraryCode from '!raw-loader!@casual-simulation/aux-custom-portals/dist/esbuild/uuid.js';
 import { addDefinitionsForLibrary } from '../../MonacoHelpers';
+import { hasValue } from '@casual-simulation/aux-common/bots/BotCalculations';
 
 @Component({
     components: {
@@ -114,7 +115,7 @@ export default class CustomPortals extends Vue {
             sim.portals.portalsUpdated
                 .pipe(
                     concatMap((u) => u),
-                    tap((update) => this._onPortalUpated(sim, update))
+                    tap((update) => this._onPortalUpdated(sim, update))
                 )
                 .subscribe()
         );
@@ -133,27 +134,31 @@ export default class CustomPortals extends Vue {
         sim: BrowserSimulation,
         portal: PortalData
     ): void {
-        this.portals.push({
-            simulationId: sim.id,
-            portalId: portal.id,
-            source: portal.source,
-            error: portal.error,
-            ports: portal.ports,
-            style: {},
-        });
+        if (hasValue(portal.source)) {
+            this.portals.push({
+                simulationId: sim.id,
+                portalId: portal.id,
+                source: portal.source,
+                error: portal.error,
+                ports: portal.ports,
+                style: {},
+            });
+        }
     }
 
-    private _onPortalUpated(
+    private _onPortalUpdated(
         sim: BrowserSimulation,
         update: PortalUpdate
     ): void {
         const portal = this.portals.find(
             (p) => p.portalId === update.portal.id
         );
-        portal.source = update.portal.source;
-        portal.style = update.portal.style;
-        portal.error = update.portal.error;
-        portal.ports = update.portal.ports;
+        if (portal) {
+            portal.source = update.portal.source;
+            portal.style = update.portal.style;
+            portal.error = update.portal.error;
+            portal.ports = update.portal.ports;
+        }
     }
 }
 
