@@ -1,4 +1,8 @@
-import { Simulation, LoginManager } from '@casual-simulation/aux-vm';
+import {
+    Simulation,
+    LoginManager,
+    PortalManager,
+} from '@casual-simulation/aux-vm';
 
 /**
  * Defines an interface for simulations that interface with remote servers.
@@ -8,4 +12,9 @@ export interface RemoteSimulation extends Simulation {
      * Gets the login manager.
      */
     login: LoginManager;
+
+    /**
+     * Gets the portal manager.
+     */
+    portals: PortalManager;
 }
